+++ src/manage/mod.rs
@@ -9,7 +9,7 @@ mod event_type;
 mod group;
 mod subscribe;
 
-pub use aio::{EnterAIO, EnterAioEvent};
+pub use aio::EnterAioEvent;
 pub use c2c::{C2CFriendData, C2CManageEvent};
 pub use event_type::ManageEventType;
 pub use group::GroupManageEvent;